requirements.txt
pyzstd
python-dotenv
scipy
